backend/requirements.txt
# Core Django
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-filter==24.2

# Auth
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary>=2.9.9

# Supabase
supabase==2.5.3

# Celery + Redis
celery==5.4.0
redis==5.0.8
django-celery-beat==2.6.0

# Channels (WebSocket)
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2

# Config
python-decouple==3.8

# Static files
whitenoise==6.7.0

# Utils
Pillow==10.4.0
python-slugify==8.0.4

# Dev
ipython==8.26.0
django-extensions==3.2.3
